Adjusted script placement and added logic
Adjusted the placement of script so that html would load before trying to run our javascript code.
creating logic for the RPG. Called all the variables needed. create functionality for some of the buttons, added locations to explore, added the ability to buy and sell weapons monsters objects for battle sections.
started working on combat.